% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inactivation}
\alias{Inactivation}
\title{Some inactivation experiments}
\format{A data frame with 75 rows and 5 variables:
\describe{
  \item{Date}{Date of the experiment}
  \item{Experimenter}{Name of the experimenter}
  \item{Timepoint}{Elapsed time}
  \item{log_cfu}{Log-microbial count}
  \item{Strain}{Microbial strain}
}}
\usage{
Inactivation
}
\description{
A dataset containing some inactivation experiments.
}
\keyword{datasets}
