% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inactivation_75K}
\alias{Inactivation_75K}
\title{A larger set of inactivation experiments}
\format{A data frame with 75000 rows and 5 variables:
\describe{
  \item{Date}{Date of the experiment}
  \item{Experimenter}{Name of the experimenter}
  \item{Timepoint}{Elapsed time}
  \item{log_cfu}{Log-microbial count}
  \item{Strain}{Microbial strain}
}}
\usage{
Inactivation_75K
}
\description{
A dataset containing more inactivation experiments.
}
\keyword{datasets}
