% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inactivation_course_filtered}
\alias{Inactivation_course_filtered}
\title{Filtered dataset of inactivation experiments for the ggplot2 lesson}
\format{A data frame with 750 rows and 6 variables:
\describe{
  \item{Date}{Date of the experiment}
  \item{Experimenter}{Name of the experimenter}
  \item{Timepoint}{Elapsed time}
  \item{log_cfu}{Log-microbial count}
  \item{Strain}{Microbial strain}
  \item{Supervisor}{Name of the tyrant}
}}
\usage{
Inactivation_course_filtered
}
\description{
A dataset containing some inactivation experiments that has been filtered
already.
}
\keyword{datasets}
